requirements.txt
streamlit
openai
pillow
requests
numpy
streamlit-drawable-canvas
supabase